requirements.txt
cx_Freeze==6.14.9
tk==0.1.0
opencv-python==4.6.0.66
mysql-connector-python==8.0.29
pillow==9.3.0
numpy==1.23.4
mediapipe==0.8.11